feat(workbench-sdk): W7a SDK 真核与 AiPanel 正式出口
将 types/registry-store/resolve-prompt/bootstrap 迁入 SDK；
提供 @sailfish/workbench-sdk/ai-panel 供岗位台引用同款对话
（实现仍在 desktop）。assistant/companion/sample 改走包名。
