+++ packages/workbench-assistant/package.json
@@ -16,6 +16,7 @@
     "vue": "^3.4.0"
   },
   "dependencies": {
-    "@sailfish/shared-types": "*"
+    "@sailfish/shared-types": "*",
+    "@sailfish/workbench-sdk": "*"
   }
 }
