+++ packages/workbench-sdk/package.json
@@ -2,14 +2,19 @@
   "name": "@sailfish/workbench-sdk",
   "version": "0.1.0",
   "private": true,
-  "description": "Workbench SDK：descriptor / 注册表 / prompt 解析（骨架）",
+  "description": "Workbench SDK：types / 注册表 / prompt / bootstrap / AiPanel 出口",
   "type": "module",
   "main": "./src/index.ts",
   "types": "./src/index.ts",
   "exports": {
-    ".": "./src/index.ts"
+    ".": "./src/index.ts",
+    "./ai-panel": "./src/ai-panel.ts",
+    "./types": "./src/types.ts"
   },
   "peerDependencies": {
-    "vue": "^3.0.0"
+    "vue": "^3.4.0"
+  },
+  "dependencies": {
+    "@sailfish/shared-types": "*"
   }
 }
