+++ packages/workbench-sdk/src/index.ts
@@ -1,23 +1,27 @@
 /**
- * @sailfish/workbench-sdk（半成品）
+ * @sailfish/workbench-sdk
  *
- * 仍 re-export 桌面内 `src/workbench` 核心 API（types / registry-store / bootstrap）。
- * 内置台 descriptor 已分别在 `@sailfish/workbench-*`；SDK 真抽待 types 迁入后再去 re-export。
+ * 工作台 SDK：类型 / 注册表 / prompt 解析 / bootstrap / AiPanel 正式出口。
+ * AiPanel 实现仍在 desktop（依赖未解耦）；岗位定制靠 descriptor.agentPrompt/skills/mcpServers。
  */
+export type {
+  WorkbenchKind,
+  WorkbenchDescriptor,
+  WorkbenchAgentPromptTab,
+  RegionSpec,
+  RegionRole,
+  RegionSide,
+} from './types'
+
 export {
   registerWorkbench,
   getWorkbenchDescriptor,
   listWorkbenchDescriptors,
-  resolveWorkbenchKind,
-  resolveWorkbenchRenderer,
-  isWorkbenchAvailable,
-  resolveWorkbenchAgentPrompt,
-  bootstrapWorkbenchCapabilities,
-} from '../../../src/workbench/index'
+  clearWorkbenchRegistryForTests,
+} from './registry-store'
 
-export type {
-  WorkbenchDescriptor,
-  WorkbenchKind,
-  RegionSpec,
-  WorkbenchAgentPromptTab,
-} from '../../../src/workbench/index'
+export { resolveWorkbenchAgentPrompt } from './resolve-workbench-agent-prompt'
+export { bootstrapWorkbenchCapabilities } from './bootstrap'
+export type { WorkbenchBootstrapResult } from './bootstrap'
+
+/** 同款对话请从 `@sailfish/workbench-sdk/ai-panel` 导入，避免主入口拉入 Vue 巨石 */
