+++ packages/workbench-ssh/package.json
@@ -13,5 +13,8 @@
   },
   "peerDependencies": {
     "vue": "^3.4.0"
+  },
+  "dependencies": {
+    "@sailfish/workbench-sdk": "*"
   }
 }
