+++ src/workbench/types.ts
@@ -1,106 +1,9 @@
-/**
- * Workbench（工作台）模型类型定义
- *
- * 一个 Tab 就是一个工作台实例。工作台 = 一组具名区域（Region）的固定组合：
- * 一个常驻锚点区 + 若干可显隐的辅助区，区域之间可拖分隔条调比例。
- *
- * 设计原则（详见会话纪要）：
- * - 组合 + 贡献，不是开关 + 枚举：工作台由区域拼成，不用一堆 hasXxx 能力位描述。
- * - 工具以叠加为主、允许覆盖：核心工具是基线，工作台按需贡献/覆盖专用工具（后端 mode 不受影响）。
- * - 布局是固定模板：区域只显隐 / 拖比例，不带焦点高亮、不带递归分割（那是终端 panel 私有的 SplitPane）。
- *
- * 渲染分两种方式：
- * - 声明式区域（regions）：交给通用 WorkbenchShell 渲染（助手、未来的浏览器等）。
- * - 自定义渲染器（renderer，逃生口）：工作台有特殊 chrome 时直接用专属组件渲染
- *   （终端工作台因 Terminal 实例 Teleport 保命池而特殊，走此路）。
- */
-import type { Component } from 'vue'
-import type { McpServerConfig, TerminalType } from '@shared/types'
-
-/**
- * 工作台类型 —— `TerminalType` 的超集。
- *
- * 多数工作台与终端类型一一对应（'local' / 'ssh' / 'assistant'），直接复用 `TerminalType`
- * 作为数据源，避免重复枚举。但**工作台外观可以独立于终端类型**：同一 `tab.type='assistant'`
- * 可映射到不同工作台（普通助手 `assistant` vs 联络 `companion`）。这类「同 type 不同工作台」
- * 的成员单独列在此处。
- *
- * `(string & {})` 保留内置 kind 的补全，同时允许业务/OEM 自定义 kind（如 `sample`）。
- *
- * 关键约定：`tab.type` 不等于 `WorkbenchKind`，二者通过 `resolveWorkbenchKind(tab)`（registry.ts）
- * 映射。新增独立工作台后必须配套该映射，否则 `tab.type` 永远查不到它。
- */
-export type WorkbenchKind = TerminalType | 'companion' | (string & {})
-
-
-/** 区域角色：anchor=常驻锚点区；toggle=可显隐辅助区 */
-export type RegionRole = 'anchor' | 'toggle'
-
-/** 区域所在侧（仅 toggle 区域有意义） */
-export type RegionSide = 'left' | 'right'
-
-/**
- * 区域声明。
- *
- * 注：当前内置工作台均通过 `renderer` 自定义组件渲染，`regions` 暂作意图声明 /
- * 未来声明式工作台使用。新增「无特殊 chrome」的工作台时，可只填 regions 交给 WorkbenchShell。
- */
-export interface RegionSpec {
-  /** 区域标识（工作台内唯一） */
-  id: string
-  /** 角色：锚点常驻 / 辅助可隐 */
-  role: RegionRole
-  /** toggle 区域所在侧，默认 'right' */
-  side?: RegionSide
-  /** 辅助区默认是否可见，默认 false */
-  defaultVisible?: boolean
-  /** 是否允许拖分隔条调尺寸，默认 true */
-  resizable?: boolean
-}
-
-/**
- * 工作台描述。一种工作台一条，集中登记在 registry，供查表分发。
- */
-export interface WorkbenchDescriptor {
-  /** 工作台类型 */
-  kind: WorkbenchKind
-  /**
-   * 自定义渲染器（逃生口）。提供则直接用它渲染整个工作台，忽略 regions。
-   * 渲染器组件统一接收 props：`{ tab: TerminalTab; isActive: boolean }`。
-   */
-  renderer?: Component
-  /**
-   * 声明式区域。无 renderer 时由通用 WorkbenchShell 按此渲染。
-   * 当前内置工作台都走 renderer，此字段为未来声明式工作台预留。
-   */
-  regions?: RegionSpec[]
-  /** 是否在 Steam 构建中可用（助手类工作台在 Steam 版不提供） */
-  availableInSteam?: boolean
-
-  // ===== 一站式打包：业务工作台带的全栈能力 =====
-
-  /** 本工作台依赖的技能 ID（desktop 启动时校验 / 装配） */
-  skills?: string[]
-  /** 本工作台自带的 MCP 配置（desktop 启动时 McpService.connect） */
-  mcpServers?: McpServerConfig[]
-  /**
-   * 工作台 Agent system prompt 片段。
-   * 字符串恒注入；函数可按 tab 决定是否注入（如 assistant 排除远程会话）。
-   */
-  agentPrompt?: string | ((tab: WorkbenchAgentPromptTab) => string | undefined)
-  /**
-   * （预留）岗位级 Agent 策略：记忆 / 召回 / 执行松紧等。
-   */
-  agentPolicy?: {
-    memory?: boolean
-    recall?: boolean
-    executionMode?: string
-  }
-}
-
-/** agentPrompt 函数入参（与 resolve-workbench-agent-prompt 对齐） */
-export interface WorkbenchAgentPromptTab {
-  type: string
-  isRemote?: boolean
-  remoteChannel?: string
-}
+/** @deprecated 兼容 re-export；真相源 `@sailfish/workbench-sdk` */
+export type {
+  WorkbenchKind,
+  WorkbenchDescriptor,
+  WorkbenchAgentPromptTab,
+  RegionSpec,
+  RegionRole,
+  RegionSide,
+} from '@sailfish/workbench-sdk'
