+++ tsconfig.json
@@ -23,6 +23,7 @@
       "@shared/*": ["./shared/*"],
       "@sailfish/shared-types": ["./packages/shared-types/src/index.ts"],
       "@sailfish/workbench-sdk": ["./packages/workbench-sdk/src/index.ts"],
+      "@sailfish/workbench-sdk/*": ["./packages/workbench-sdk/src/*"],
       "@sailfish/workbench-assistant": ["./packages/workbench-assistant/src/index.ts"],
       "@sailfish/workbench-assistant/*": ["./packages/workbench-assistant/src/*"],
       "@sailfish/workbench-sample": ["./packages/workbench-sample/src/index.ts"],
@@ -41,6 +42,7 @@
     "src/**/*.vue",
     "shared/**/*.ts",
     "packages/shared-types/src/**/*.ts",
+    "packages/workbench-sdk/src/**/*.ts",
     "packages/workbench-assistant/src/**/*.ts",
     "packages/workbench-assistant/src/**/*.vue",
     "packages/workbench-sample/src/**/*.ts",
