+++ vitest.config.ts
@@ -20,6 +20,7 @@ export default defineConfig({
       '@shared/types': resolve(__dirname, 'packages/shared-types/src'),
       '@shared': resolve(__dirname, 'shared'),
       '@sailfish/shared-types': resolve(__dirname, 'packages/shared-types/src/index.ts'),
+      '@sailfish/workbench-sdk': resolve(__dirname, 'packages/workbench-sdk/src'),
       '@sailfish/workbench-assistant': resolve(__dirname, 'packages/workbench-assistant/src'),
       '@sailfish/workbench-sample': resolve(__dirname, 'packages/workbench-sample/src'),
       '@sailfish/workbench-local': resolve(__dirname, 'packages/workbench-local/src'),
